[TM-46] fix: Refactored Moy Sklad controller
