+++ src/controllers/moySkladMarketController.ts
@@ -1,38 +1,27 @@
-import { Request, Response } from 'express';
+import { toBooleanOrUndefined, toNumberOrUndefined } from '../utils/query';
 import { ProductFeedService } from '../services/ProductFeedService';
 import { ListParams } from '../MoySkladApi/MoySkladTypes';
-
-const toNumberOrUndefined = (value: unknown): number | undefined => {
-    if (value === undefined) return undefined;
-
-    const numeric = Number(value);
-
-    return Number.isFinite(numeric) ? numeric : undefined;
-};
-
-const toBooleanOrUndefined = (value: unknown): boolean | undefined => {
-    if (value === undefined) return undefined;
-
-    if (value === 'true' || value === true) return true;
-
-    if (value === 'false' || value === false) return false;
-
-    return undefined;
-};
+import { Request, Response } from 'express';
+import {
+    DEFAULT_LIMIT,
+    DEFAULT_INCLUDE_IMAGES,
+    DEFAULT_OFFSET,
+} from '../utils/constants';
 
 export class MoySkladMarketController {
     constructor(private readonly feed: ProductFeedService) {}
 
     fetchMarketProducts = async (request: Request, response: Response) => {
         const params: ListParams & { reviewsLimit?: number } = {
-            limit: toNumberOrUndefined(request.query.limit) ?? 50,
-            offset: toNumberOrUndefined(request.query.offset) ?? 0,
+            limit: toNumberOrUndefined(request.query.limit) ?? DEFAULT_LIMIT,
+            offset: toNumberOrUndefined(request.query.offset) ?? DEFAULT_OFFSET,
             search:
                 typeof request.query.search === 'string'
                     ? request.query.search
                     : undefined,
             includeImages:
-                toBooleanOrUndefined(request.query.includeImages) ?? true,
+                toBooleanOrUndefined(request.query.includeImages) ??
+                DEFAULT_INCLUDE_IMAGES,
             onlyActive: toBooleanOrUndefined(request.query.onlyActive),
             reviewsLimit: toNumberOrUndefined(request.query.reviewsLimit),
         };
